+++ Unsafe/unsafe/SynchronizedCounter.java
@@ -1,17 +1,17 @@
 package unsafe;
 
 public final class SynchronizedCounter implements Counter {
-  private int i;
+  private long i;
 
   @Override
-  public int increment() {
+  public long increment() {
     synchronized (this) {
       return ++i;
     }
   }
 
   @Override
-  public int get() {
+  public long get() {
     synchronized (this) {
       return i;
     }
